MM-18340: restore dialog's store.subscribe
https://github.com/mattermost/mattermost-webapp/pull/3493 unintentionally removed the import of `actions/integration_actions`, which defined a side effect necessary to the correct functioning of interactive dialogs.

I've relocated this somewhere that is more intentionally imported, and commented details as to the issues with this code block. I'm not explicitly fixing the issues right now, and I'm not sure if I can unit test this, so I'm just restoring the import for now.
