+++ actions/integration_actions.jsx
@@ -6,11 +6,6 @@ import * as IntegrationActions from 'mattermost-redux/actions/integrations';
 import {getProfilesByIds} from 'mattermost-redux/actions/users';
 import {getUser} from 'mattermost-redux/selectors/entities/users';
 
-import {ModalIdentifiers} from 'utils/constants';
-import {openModal} from 'actions/views/modals';
-import InteractiveDialog from 'components/interactive_dialog';
-import store from 'stores/redux_store.jsx';
-
 const DEFAULT_PAGE_SIZE = 100;
 
 export function loadIncomingHooksAndProfilesForTeam(teamId, page = 0, perPage = DEFAULT_PAGE_SIZE) {
@@ -144,22 +139,3 @@ export function getYoutubeVideoInfo(googleKey, videoId, success, error) {
             return success(res.body);
         });
 }
-
-let previousTriggerId = '';
-store.subscribe(() => {
-    const state = store.getState();
-    const currentTriggerId = state.entities.integrations.dialogTriggerId;
-
-    if (currentTriggerId === previousTriggerId) {
-        return;
-    }
-
-    previousTriggerId = currentTriggerId;
-
-    const dialog = state.entities.integrations.dialog || {};
-    if (dialog.trigger_id !== currentTriggerId) {
-        return;
-    }
-
-    store.dispatch(openModal({modalId: ModalIdentifiers.INTERACTIVE_DIALOG, dialogType: InteractiveDialog}));
-});
